fix: rename AcronymPatchValidator package and add path existence check for validation
